pyproject.toml
[project]
name = "neo4j-databricks-demo"
version = "0.1.0"
description = "Demo project for Neo4j and Databricks integration"
requires-python = ">=3.10"
dependencies = [
    "databricks-sdk>=0.20.0",
    "neo4j>=5.0.0",
    "pyspark>=3.5.0",
    "python-dotenv>=1.0.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.0.0",
    # V2 agent: Native structured output with ChatDatabricks
    # Reference: https://api-docs.databricks.com/python/databricks-ai-bridge/latest/databricks_langchain.html
    "databricks-langchain>=0.1.0",
    # DSPy agent: Programmatic AI with structured output
    # Reference: https://docs.databricks.com/aws/en/generative-ai/dspy/
    # PyPI: https://pypi.org/project/dspy/ (v3.0.4 as of Nov 2025)
    "dspy>=3.0.4",
    # MLflow for DSPy tracing (optional but recommended)
    # Reference: https://docs.databricks.com/aws/en/mlflow3/genai/tracing/integrations/dspy
    "mlflow[databricks]>=3.1",
]
